+++ libs/@local/hashql/mir/src/pass/execution/island/graph/tests.rs
@@ -1,25 +1,30 @@
-//! Tests for island dependency graph construction and fetch island insertion.
+//! Tests for island dependency graph construction and requirement resolution.
+#![expect(clippy::min_ident_chars)]
 
 use alloc::alloc::Global;
+use core::assert_matches;
 
-use hashql_core::{heap::Heap, symbol::sym, r#type::environment::Environment};
+use hashql_core::{
+    graph::DirectedGraph as _, heap::Heap, id::Id as _, symbol::sym,
+    r#type::environment::Environment,
+};
 
 use crate::{
-    body::basic_block::{BasicBlockId, BasicBlockVec},
+    body::{Body, basic_block::BasicBlockVec},
     builder::body,
     intern::Interner,
     pass::execution::{
         VertexType,
         island::{
             IslandId, IslandPlacement,
-            graph::{IslandEdgeKind, IslandNode, IslandNodeId, build_island_graph},
+            graph::{IslandEdge, IslandGraph, IslandKind},
         },
         target::TargetId,
         traversal::EntityPath,
     },
 };
 
-fn make_targets(assignments: &[TargetId]) -> BasicBlockVec<TargetId, Global> {
+pub(crate) fn make_targets(assignments: &[TargetId]) -> BasicBlockVec<TargetId, Global> {
     let mut targets = BasicBlockVec::with_capacity_in(assignments.len(), Global);
     for &target in assignments {
         targets.push(target);
@@ -27,7 +32,15 @@ fn make_targets(assignments: &[TargetId]) -> BasicBlockVec<TargetId, Global> {
     targets
 }
 
-/// Single Postgres island accessing properties — no fetch island needed because the island
+pub(crate) fn build_graph(body: &Body<'_>, targets: &[TargetId]) -> IslandGraph<Global> {
+    let target_vec = make_targets(targets);
+    let islands = IslandPlacement::new().run(body, VertexType::Entity, &target_vec, Global);
+    let mut graph = IslandGraph::build_in(body, VertexType::Entity, islands, Global, Global);
+    graph.resolve(Global);
+    graph
+}
+
+/// Single Postgres island accessing properties: no fetch island needed because the island
 /// itself is on the origin backend for that path.
 #[test]
 fn single_island_no_fetch() {
@@ -45,21 +58,16 @@ fn single_island_no_fetch() {
         }
     });
 
-    let targets = make_targets(&[TargetId::Postgres]);
-    let islands = IslandPlacement::new().run(&body, VertexType::Entity, &targets, Global);
-    let graph = build_island_graph(&body, &islands, VertexType::Entity);
+    let graph = build_graph(&body, &[TargetId::Postgres]);
 
-    // One real node, no fetch islands.
-    assert_eq!(graph.nodes.len(), 1);
-    assert!(
-        matches!(graph.nodes[IslandNodeId::new(0)], IslandNode::Real(id) if id == IslandId::new(0))
-    );
-    assert!(graph.edges.is_empty());
+    assert_eq!(graph.node_count(), 1);
+    assert_matches!(graph[IslandId::new(0)].kind(), IslandKind::Exec(_));
+    assert_eq!(graph.edge_count(), 0);
 }
 
 /// Postgres island followed by Interpreter island that needs properties.
-/// Properties originate from Postgres, so the Interpreter island gets a data edge
-/// from the Postgres island — no fetch island needed.
+/// Properties originate from Postgres, so the Interpreter island gets a `DataFlow` edge
+/// from the Postgres island: no fetch island needed.
 #[test]
 fn data_edge_from_predecessor() {
     let heap = Heap::new();
@@ -80,36 +88,39 @@ fn data_edge_from_predecessor() {
         }
     });
 
-    let targets = make_targets(&[TargetId::Postgres, TargetId::Interpreter]);
-    let islands = IslandPlacement::new().run(&body, VertexType::Entity, &targets, Global);
+    let graph = build_graph(&body, &[TargetId::Postgres, TargetId::Interpreter]);
 
-    assert_eq!(islands.len(), 2);
+    // Two exec nodes, no data islands.
+    assert_eq!(graph.node_count(), 2);
 
-    let graph = build_island_graph(&body, &islands, VertexType::Entity);
+    let control_flow_count = graph
+        .iter_edges()
+        .filter(|edge| edge.data == IslandEdge::ControlFlow)
+        .count();
+    assert_eq!(control_flow_count, 1);
 
-    // Two real nodes, no fetch islands.
-    assert_eq!(graph.nodes.len(), 2);
+    let data_flow_count = graph
+        .iter_edges()
+        .filter(|edge| edge.data == IslandEdge::DataFlow)
+        .count();
+    assert_eq!(data_flow_count, 1);
 
-    // Should have a CFG edge from Postgres to Interpreter.
-    let cfg_edges: Vec<_> = graph
-        .edges
-        .iter()
-        .filter(|edge| edge.kind == IslandEdgeKind::Control)
-        .collect();
-    assert_eq!(cfg_edges.len(), 1);
-
-    // Should have a data edge carrying Properties from Postgres to Interpreter.
-    let data_edges: Vec<_> = graph
-        .edges
-        .iter()
-        .filter(|edge| edge.kind == IslandEdgeKind::Data)
-        .collect();
-    assert_eq!(data_edges.len(), 1);
-    assert!(data_edges[0].paths.contains(EntityPath::Properties));
+    // The Postgres island should provide Properties.
+    let postgres_island = (0..graph.node_count())
+        .map(IslandId::from_usize)
+        .find(|&island_id| graph[island_id].target() == TargetId::Postgres)
+        .expect("postgres island exists");
+    assert!(
+        graph[postgres_island]
+            .provides()
+            .as_entity()
+            .expect("entity vertex")
+            .contains(EntityPath::Properties)
+    );
 }
 
 /// Interpreter island needs embedding data but has no Embedding predecessor.
-/// A FetchIsland(Embedding) should be inserted.
+/// A data island for Embedding should be inserted.
 #[test]
 fn fetch_island_for_unsatisfied_requirement() {
     let heap = Heap::new();
@@ -127,30 +138,29 @@ fn fetch_island_for_unsatisfied_requirement() {
         }
     });
 
-    let targets = make_targets(&[TargetId::Interpreter]);
-    let islands = IslandPlacement::new().run(&body, VertexType::Entity, &targets, Global);
-    let graph = build_island_graph(&body, &islands, VertexType::Entity);
+    let graph = build_graph(&body, &[TargetId::Interpreter]);
 
-    // One real node plus one FetchIsland(Embedding).
-    assert_eq!(graph.nodes.len(), 2);
-    assert!(matches!(
-        graph.nodes[IslandNodeId::new(0)],
-        IslandNode::Real(_)
-    ));
+    // One exec node plus one data island for Embedding.
+    assert_eq!(graph.node_count(), 2);
 
-    let fetch_node = &graph.nodes[IslandNodeId::new(1)];
-    match fetch_node {
-        IslandNode::Fetch(fetch) => {
-            assert_eq!(fetch.target, TargetId::Embedding);
-            assert!(fetch.paths.contains(EntityPath::Vectors));
-        }
-        IslandNode::Real(_) => panic!("expected FetchIsland, got Real"),
-    }
+    let data_island = (0..graph.node_count())
+        .map(IslandId::from_usize)
+        .find(|&island_id| matches!(graph[island_id].kind(), IslandKind::Data))
+        .expect("data island exists");
+
+    assert_eq!(graph[data_island].target(), TargetId::Embedding);
+    assert!(
+        graph[data_island]
+            .provides()
+            .as_entity()
+            .expect("entity vertex")
+            .contains(EntityPath::Vectors)
+    );
 }
 
 /// Diamond CFG: Postgres branches to Interpreter and Embedding, both merge into a
 /// final Postgres island. The Embedding path is only available on one branch, so the
-/// final Postgres island needs a FetchIsland for embedding data.
+/// final Postgres island needs a data island for embedding data.
 #[test]
 fn diamond_branch_needs_fetch() {
     let heap = Heap::new();
@@ -183,75 +193,74 @@ fn diamond_branch_needs_fetch() {
     });
 
     // bb0=Postgres, bb1=Interpreter, bb2=Embedding, bb3=Postgres
-    let targets = make_targets(&[
-        TargetId::Postgres,
-        TargetId::Interpreter,
-        TargetId::Embedding,
-        TargetId::Postgres,
-    ]);
-    let islands = IslandPlacement::new().run(&body, VertexType::Entity, &targets, Global);
-    let graph = build_island_graph(&body, &islands, VertexType::Entity);
+    let graph = build_graph(
+        &body,
+        &[
+            TargetId::Postgres,
+            TargetId::Interpreter,
+            TargetId::Embedding,
+            TargetId::Postgres,
+        ],
+    );
 
-    // The final Postgres island needs Vectors. The Embedding island (bb2) only runs on
-    // one branch and doesn't dominate bb3, so a FetchIsland(Embedding) must be inserted.
-    let fetch_nodes: Vec<_> = graph
-        .nodes
-        .iter()
-        .filter(|node| matches!(node, IslandNode::Fetch(_)))
-        .collect();
+    // The Embedding island (bb2) only runs on one branch and doesn't dominate bb3,
+    // so a data island for Embedding must be inserted.
+    let has_embedding_data_island =
+        (0..graph.node_count())
+            .map(IslandId::from_usize)
+            .any(|island_id| {
+                matches!(graph[island_id].kind(), IslandKind::Data)
+                    && graph[island_id].target() == TargetId::Embedding
+            });
 
     assert!(
-        !fetch_nodes.is_empty(),
-        "expected at least one FetchIsland for embedding data"
+        has_embedding_data_island,
+        "expected a data island for Embedding"
     );
-
-    let has_embedding_fetch = fetch_nodes.iter().any(
-        |node| matches!(node, IslandNode::Fetch(fetch) if fetch.target == TargetId::Embedding),
-    );
-    assert!(has_embedding_fetch);
 }
 
-/// Schedule levels: independent fetch islands and entry islands should be at level 0.
-/// Their consumers should be at level 1 or higher.
+/// Inherits edge: when two same-target islands are in a dominator relationship,
+/// the child inherits provided paths from the parent.
 #[test]
-fn schedule_levels() {
+fn inherits_edge_same_target_dominator() {
     let heap = Heap::new();
     let interner = Interner::new(&heap);
     let env = Environment::new(&heap);
 
     let body = body!(interner, env; [graph::read::filter]@0/2 -> ? {
-        decl env: (), vertex: [Opaque sym::path::Entity; ?], val: ?;
-        @proj enc = vertex.encodings: ?,
-              vecs = enc.vectors: ?;
+        decl env: (), vertex: [Opaque sym::path::Entity; ?], val1: ?, val2: ?;
+        @proj props = vertex.properties: ?;
 
         bb0() {
-            val = load vecs;
-            return val;
+            val1 = load props;
+            goto bb1();
+        },
+        bb1() {
+            goto bb2();
+        },
+        bb2() {
+            val2 = load props;
+            return val2;
         }
     });
 
-    let targets = make_targets(&[TargetId::Interpreter]);
-    let islands = IslandPlacement::new().run(&body, VertexType::Entity, &targets, Global);
-    let graph = build_island_graph(&body, &islands, VertexType::Entity);
+    // bb0=Postgres, bb1=Interpreter, bb2=Postgres
+    // bb0 dominates bb2 (through bb1). Both are Postgres, so bb2 should inherit from bb0.
+    let graph = build_graph(
+        &body,
+        &[
+            TargetId::Postgres,
+            TargetId::Interpreter,
+            TargetId::Postgres,
+        ],
+    );
 
-    // The real island depends on the FetchIsland, so it should be at a higher level.
-    assert!(graph.schedule.len() >= 2);
-
-    let real_level = graph
-        .schedule
-        .iter()
-        .find(|sched| matches!(graph.nodes[sched.node], IslandNode::Real(_)))
-        .map(|sched| sched.level);
-    let fetch_level = graph
-        .schedule
-        .iter()
-        .find(|sched| matches!(graph.nodes[sched.node], IslandNode::Fetch(_)))
-        .map(|sched| sched.level);
-
-    if let (Some(real_level), Some(fetch_level)) = (real_level, fetch_level) {
-        assert!(
-            real_level > fetch_level,
-            "real island (level {real_level}) should be after fetch island (level {fetch_level})"
-        );
-    }
+    let inherits_count = graph
+        .iter_edges()
+        .filter(|edge| edge.data == IslandEdge::Inherits)
+        .count();
+    assert!(
+        inherits_count > 0,
+        "expected an Inherits edge between same-target dominating islands"
+    );
 }
