+++ libs/@local/hashql/mir/src/pass/execution/island/mod.rs
@@ -29,8 +29,8 @@ use crate::{
     visit::Visitor as _,
 };
 
-// pub(crate) mod graph;
-mod graph2;
+pub(crate) mod graph;
+pub(crate) mod schedule;
 #[cfg(test)]
 mod tests;
